refactor: [#100] move classes to new mod
