+++ src/nautilus_librarian/mods/filesystem/domain/directory.py
