+++ tests/test_nautilus_librarian/test_mods/test_filesystem/test_domain/test_filepath.py
@@ -1,6 +1,6 @@
-from nautilus_librarian.domain.directory import Directory
-from nautilus_librarian.domain.filename import Filename
-from nautilus_librarian.domain.filepath import Filepath
+from nautilus_librarian.mods.filesystem.domain.directory import Directory
+from nautilus_librarian.mods.filesystem.domain.filename import Filename
+from nautilus_librarian.mods.filesystem.domain.filepath import Filepath
 
 
 def it_could_be_compared_to_other_filepath():
